### 2021-01-08 #421 @mskyttner
aquamapsdata - Curated Data From AquaMaps.Org
Submitting Author Name:  Markus Skyttner
Submitting Author Github Handle: @mskyttner
Repository: https://github.com/raquamaps/aquamapsdata
Version submitted: 0.1.6

Version accepted: TBD

---

```
Package: aquamapsdata
Title: Curated Data From AquaMaps.Org
Version: 0.1.6
  person("Markus", "Skyttner", email = "markus.skyttner@nrm.se", role = c("aut", "cre")),
  person("Christina", "Garilao", role = c("ctb")))
Maintainer: Markus Skyttner <markus.skyttner@nrm.se>
Description: This package provides access to curated data from aquamaps.org published in 2020.
    It does not embed the data statically into the package, instead it attempts to download an SQLite database from a remote location.
    It has a function called download_db() which retrieves the data published by aquamaps.org as an SQLite database.
    It also has functions for accessing the data. The data is expected to be updated on the remote location approximately once per year.
License: file LICENSE
    dbplyr,
    RSQLite,
    R.utils,
    stringi,
    DBI,
    RMySQL,
    duckdb,
    leaflet,
    rappdirs,
    raster,
    rcrypt,
    sp,
    curl
    httr,
    readxl,
    xml2,
    DT,
    covr
SystemRequirements:
    rgdal

```

AquaMaps generates standardized computer-generated and fairly reliable large scale predictions of marine and freshwater species. The package provides functions for primarily accessing (and secondarily managing) curated data from aquamaps.org. It does not embed the data statically into the package itself; instead a function allows for a download of a compressed archive from archive.org, containing the full dataset in a single around 10+ GB large SQLite database file (close to 2 GB when compressed).

The primary target audience is researchers studying ecology for aquatic species such as fish populations, when doing studies in ecology (federal agencies, academic researchers, industry users). Another audience consists of the people at aquamaps.org that manage the data. They have needs for workflows that allow them to expose the data using an R package and therefore need some functions to assist with converting and syncing data, making an extract (by minifying/slicing the data) and in general getting it in shape for being used in the package (including uploading data to archive.org, with approximately yearly frequency).

I haven't found any, but in the last couple of years I have looked mostly in a narrow fashion with regards to packages that provide data specifically from aquamaps.org. If widening the scope, the rfishbase and taxizedb packages are partly overlapping since these packages too provide for example for retrieving taxonomic information about aquatic species, but they're have a different scope or focus, since aquamapsdata is focused on providing spatial raster data for native habitats as predicted by the aquamaps algorithm.

N/A
