### 2021-01-12 #421 @melvidoni
Hello @mskyttner, thanks for the submission. Before proceeding, the Editors have some questions for you.

1.  We are not sure whether this package helps access data or distributes data. There is the impression the authors pre-prepared some dataset from the website they mention.

2. Are there no other more dynamic ways to get the data? Especially as in the future their prepared dataset might get stale.

3. Also, some editors had troubles accessing aquamaps.org, as they got a security alert. Can you double-check that the provided link is correct?
