### 2021-01-12 #421 @mskyttner
Hello @melvidoni,

1. Data for curated aquamaps as exposed by https://www.aquamaps.org is stored in a mysql database that powers that website, some more background about the data can be found [here](https://www.gbif.org/tool/81356/aquamaps-predicted-range-maps-for-aquatic-species). The aquamapsdata R package is a little bit like taxizedb in the sense that it takes almost all data in this database and makes it accessible from R to researchers, but it does not access the mysql database directly, instead it uses a compressed archive variant in SQLite format that gets downloaded from archive.org. So yes, it does help with accessing the data. But there is also a set of functions in the package to support distribution ("data management") that can be used by aquamaps staff themselves to generate this archive, ie for converting the whole mysql database into a SQLite database (with indexes and full text search capabilities) and to push a compressed archive to archive.org. So yes, it helps with distributing the data, too. These functions for "data management/distribution" is what is pulling down the test coverage stats. To summarize: yes, there are functions in the package for these different audiences - for data access (researchers, consumers of the data) and for distributing the data (aquamaps staff and their workflows as producers of the data). Can you clarify "There is the impression the authors pre-prepared some dataset from the website they mention.", which authors and website are you referring to? Some more info on the data scope is available here: https://raquamaps.github.io/aquamapsdata/articles/intro.html#scope
2. One could web scrape the aquamaps.org website. As in the archived https://github.com/ropensci-archive/zissou R package. But my understanding is that the aquamaps.org staff doesn't encourage that. There is no API yet, as far as I know.
3. Thanks for bringing this up, I checked https://www.ssllabs.com/ssltest/analyze.html?d=www.aquamaps.org and that seems to be OK, so it looks like I could change those links to https://www.aquamaps.org, but before doing that I'll first check with the aquamaps.org staff if they can also add aquamaps.org as a valid SNI name (it seems to resolve with a warning if skipping the www part, with the message "This server could not prove that it is aquamaps.org; its security certificate is from www.fishbase.de"; possibly this is a configuration that can be adjusted).
