### 2021-01-17 #421 @melvidoni
Thank you for the detailed response @mskyttner. Could you please clarify who updates the "compressed archive variant in SQLite format that gets downloaded from archive.org" and how often it is done?
