### 2021-01-18 #421 @mskyttner
Yes @melvidoni, this is done by AquaMaps staff using the "data management" functions in the package and scripts like those in [data-raw](https://github.com/raquamaps/aquamapsdata/blob/master/data-raw/am_upload_ia.R), a workflow which they have requested I developed support for in the package, in order to be able to do those updates more frequently, when data is changed or appended at the source. Historically the update frequency has been sporadic, with yearly updates. I have developed the package on basis of input from the AquaMaps team and will ask the staff there what their planned update frequency is and report back here going forward. Also, the database structure in the export / data distribution is "cleaned up" with regards to field names, scope/filters in order to be able to distribute the relevant data using a new more easily automatable workflow.
