### 2021-01-23 #421 @maurolepore
@mskyttner thanks for your submission! It's my pleasure to be the handling editor.

Here are my editor checks ([guidelines](https://devguide.ropensci.org/editorguide.html#editorchecklist)). I'll start exploring potential reviewers but I would like to assign them only after you meet or justify against my requests (see section "Requests" below). This is to ensure the reviewers focus in the most interesting aspects of your work, and ultimately to help you maximize the quality of your package. Let me know when you are ready.

I also added a section "Observations". This is to support my requests and to guide you in finding a solution or argument against my requests.

Please let me know if you have any question or concern.

## Key editor checks

### Requests

FIT AND OVERLAP

-   [ ] If still relevant, please also argue explicitly how this package   fits in the category “geospatial data”.

-   [ ] Please refer to the  [guidelines](https://devguide.ropensci.org/policies.html#overlap)
    and discuss the potential overlap with other packages.

INSTALLATION

-   [ ] Do users need the system dependency `gpg`? If so, please add it   to DESCRIPTION.

INTEGRITY

-   [ ] Does DESCRIPTION need “rgdal” to include under Imports or   Suggests?

-   [ ] Please address (or justify against) the recommendations of  `goodpractice::gp()`.

METADATA

-   [ ] Please fix problematic URLs in README and DESCRIPTION.

-   [ ] Please review possibly misspelled words.

TESTS

-   [ ] Please thrive to increase test coverage or use [exclusion comments](https://github.com/r-lib/covr#exclusion-comments) if   appropriate.

CI

-   [ ] Please expand the configuration for continuous integration to   check the package with as many platforms and versions of R as   possible. DESCRIPTION claims to depend on R &gt;= 3.0 so it would be  nice to confirm this claim with automated checks.

STYLE

-   [ ] Please ensure source code is styled in a way that facilitates   reviews, e.g. via `styler::style_pkg()` and   `usethis::use_tidy_description()`.

### Observations

FIT AND OVERLAP

-   Here I see an argument for how the package fits in “data retrieval”   but not in “geospatial data”.

-   In README I see “A similar package which also provides the aquamaps   algorithm is available at <https://github.com/raquamaps/raquamaps>”;   but overlap is not discussed here.

INSTALLATION

-   Local install on my personal computer and personal library succeeds.

-   Local install on a container from rocker/verse fails with:

``` r
install(dependencies = TRUE, build_vignettes = TRUE)

Error: package or namespace load failed for ‘rcrypt’:
 .onLoad failed in loadNamespace() for 'rcrypt', details:
  call: fun(libname, pkgname)
  error: Please install GPG first or check if GPG can be run from the command line.
      The rcrypt package is just an interface to GPG. See https://gnupg.org/ for
      installation guidelines.
```

-   Installation from GitHub on my personal computer succeeds; but on a container from rocker/verse fails as above.

INTEGRITY

-   `devtools::check()` failed with a message suggesting a vignette   needs it, and it seems to not install with the package.

-   Also I experienced problems with codetools, but it might be a an odd  problem specific to my system.

``` r
Error in .doLoadActions(where, attach) :
  error in load action .__A__.1 for package raster: loadModule(module = "spmod", what = TRUE, env = ns, loadNow = TRUE): Unable to load module "spmod": package ‘codetools’ was installed before R 4.0.0: please re-install it
Calls: <Anonymous> ... asNamespace -> loadNamespace -> <Anonymous> -> .doLoadActions
```

<summary>
Details
</summary>

``` r
devtools::check()
#> Updating aquamapsdata documentation
#> Loading aquamapsdata
#>
#>   _.  _.      _. ._ _   _. ._   _  _|  _. _|_  _.
#>  (_| (_| |_| (_| | | | (_| |_) _> (_| (_|  |_ (_|
#>        |                   |
#> Writing NAMESPACE
#> Writing NAMESPACE
#> ── Building ──────────────────────────────────────────────────── aquamapsdata ──
#> Setting env vars:
#> ● CFLAGS    : -Wall -pedantic
#> ● CXXFLAGS  : -Wall -pedantic
#> ● CXX11FLAGS: -Wall -pedantic
#> ────────────────────────────────────────────────────────────────────────────────
#>      checking for file ‘/home/mauro/git/aquamapsdata/DESCRIPTION’ ...  ✓  checking for file ‘/home/mauro/git/aquamapsdata/DESCRIPTION’
#>   ─  preparing ‘aquamapsdata’:
#>      checking DESCRIPTION meta-information ...  ✓  checking DESCRIPTION meta-information
#>   ─  checking for LF line-endings in source and make files and shell scripts
#>   ─  checking for empty or unneeded directories
#>   ─  building ‘aquamapsdata_0.1.6.tar.gz’
#>
#> ── Checking ──────────────────────────────────────────────────── aquamapsdata ──
#> Setting env vars:
#> ● _R_CHECK_CRAN_INCOMING_USE_ASPELL_: TRUE
#> ● _R_CHECK_CRAN_INCOMING_REMOTE_    : FALSE
#> ● _R_CHECK_CRAN_INCOMING_           : FALSE
#> ● _R_CHECK_FORCE_SUGGESTS_          : FALSE
#> ● NOT_CRAN                          : true
#> ── R CMD check ─────────────────────────────────────────────────────────────────
#> * using log directory ‘/tmp/Rtmp23ztZ6/aquamapsdata.Rcheck’
#> * using R version 4.0.3 (2020-10-10)
#> * using platform: x86_64-pc-linux-gnu (64-bit)
#> * using session charset: UTF-8
#> * using options ‘--no-manual --as-cran’
#> * checking for file ‘aquamapsdata/DESCRIPTION’ ... OK
#> * checking extension type ... Package
#> * this is package ‘aquamapsdata’ version ‘0.1.6’
#> * package encoding: UTF-8
#> * checking package namespace information ... OK
#> * checking package dependencies ... OK
#> * checking if this is a source package ... OK
#> * checking if there is a namespace ... OK
#> * checking for executable files ... OK
#> * checking for hidden files and directories ... OK
#> * checking for portable file names ... OK
#> * checking for sufficient/correct file permissions ... OK
#> * checking serialization versions ... OK
#> * checking whether package ‘aquamapsdata’ can be installed ... ERROR
#> Installation failed.
#> See ‘/tmp/Rtmp23ztZ6/aquamapsdata.Rcheck/00install.out’ for details.
#> * DONE
#>
#> Status: 1 ERROR
#> See
#>   ‘/tmp/Rtmp23ztZ6/aquamapsdata.Rcheck/00check.log’
#> for details.
#> ── R CMD check results ───────────────────────────────── aquamapsdata 0.1.6 ────
#> Duration: 28.4s
#>
#> > checking whether package ‘aquamapsdata’ can be installed ... ERROR
#>   See below...
#>
#> ── Install failure ─────────────────────────────────────────────────────────────
#>
#> * installing *source* package ‘aquamapsdata’ ...
#> ** using staged installation
#> ** R
#> ** data
#> *** moving datasets to lazyload DB
#> ** exec
#> ** inst
#> ** byte-compile and prepare package for lazy loading
#> Error in .doLoadActions(where, attach) :
#>   error in load action .__A__.1 for package raster: loadModule(module = "spmod", what = TRUE, env = ns, loadNow = TRUE): Unable to load module "spmod": package ‘codetools’ was installed before R 4.0.0: please re-install it
#> Calls: <Anonymous> ... asNamespace -> loadNamespace -> <Anonymous> -> .doLoadActions
#> Execution halted
#> ERROR: lazy loading failed for package ‘aquamapsdata’
#> * removing ‘/tmp/Rtmp23ztZ6/aquamapsdata.Rcheck/aquamapsdata’
#>
#> 1 error x | 0 warnings ✓ | 0 notes ✓
#> Error: R CMD check found ERRORs
```

-   All tests pass.

``` r
devtools::test()
#> Loading aquamapsdata
#>
#>   _.  _.      _. ._ _   _. ._   _  _|  _. _|_  _.
#>  (_| (_| |_| (_| | | | (_| |_) _> (_| (_|  |_ (_|
#>        |                   |
#> Testing aquamapsdata
#> ✓ |  OK F W S | Context
#> ⠏ |   0       | csquare                                                         ⠏ |   0       | Csquare Conversions                                             ✓ |  12       | Csquare Conversions
#> ⠏ |   0       | name_search                                                     ⠏ |   0       | Name Search                                                     ⠧ |   8       | Name Search                                                     ✓ |   9       | Name Search [0.1 s]
#>
#> ══ Results ═════════════════════════════════════════════════════════════════════
#> Duration: 0.2 s
#>
#> [ FAIL 0 | WARN 0 | SKIP 0 | PASS 21 ]
```

-   I see some good practices are not observed:

<summary>
Details
</summary>

``` r
goodpractice::gp()
#> Preparing: covr
#> Preparing: cyclocomp
#> Warning in MYPREPS[[prep]](state, quiet = quiet): Prep step for cyclomatic
#> complexity failed.
#> Preparing: description
#> Preparing: lintr
#> Preparing: namespace
#> Preparing: rcmdcheck
#> Warning in MYPREPS[[prep]](state, quiet = quiet): Prep step for rcmdcheck
#> failed.
#> ── GP aquamapsdata ─────────────────────────────────────────────────────────────
#>
#> It is good practice to
#>
#>   ✖ write unit tests for all functions, and all package code in
#>     general. 25% of code lines are covered by test cases.
#>
#>     R/aquamapsdata-package.R:89:NA
#>     R/aquamapsdata-package.R:90:NA
#>     R/aquamapsdata-package.R:91:NA
#>     R/aquamapsdata-package.R:92:NA
#>     R/aquamapsdata-package.R:93:NA
#>     ... and 427 more lines
#>
#>   ✖ omit "Date" in DESCRIPTION. It is not required and it gets invalid
#>     quite often. A build date will be added to the package when you
#>     perform `R CMD build` on it.
#>   ✖ avoid long code lines, it is bad for readability. Also, many people
#>     prefer editor windows that are about 80 characters wide. Try make
#>     your lines shorter than 80 characters
#>
#>     R/spatial.R:57:1
#>     R/spatial.R:70:1
#>     R/spatial.R:83:1
#>
#>   ✖ avoid sapply(), it is not type safe. It might return a vector, or a
#>     list, depending on the input data. Consider using vapply() instead.
#>
#>     R/csquare.R:129:11
#>     R/csquare.R:130:11
#>     R/csquare.R:131:11
#>
#>   ✖ not import packages as a whole, as this can cause name clashes
#>     between the imported packages. Instead, import only the specific
#>     functions you need.
#> ────────────────────────────────────────────────────────────────────────────────
```

METADATA

-   I see some problematic URLs

``` r
> urlchecker::url_check()
! Warning: README.md:84:19 Moved
Unported License](http://creativecommons.org/licenses/by-nc/3.0/):
                  ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
                  https://creativecommons.org/licenses/by-nc/3.0/
x Error: README.md:32:2 Error: SSL: no alternative certificate subject name matches target host name 'aquamaps.org'
<https://aquamaps.org>, but the `aquamapsdata` package offers an a
 ^~~~~~~~~~~~~~~~~~~~
x Error: README.md:71:43 Error: SSL: no alternative certificate subject name matches target host name 'aquamaps.org'
`aquamapsdata` provides data output from <https://aquamaps.org>; when
                                          ^~~~~~~~~~~~~~~~~~~~
x Error: man/am_meta.Rd:18:7 Error: SSL: no alternative certificate subject name matches target host name 'aquamaps.org'
\href{https://aquamaps.org/}{AquaMaps}
      ^~~~~~~~~~~~~~~~~~~~~
```

SPELLCHECK

* I see unknown or possibly misspelled words:

<summary>
Details
</summary>

``` r
spelling::spell_check_package()
#> DESCRIPTION does not contain 'Language' field. Defaulting to 'en-US'.
#>   WORD                    FOUND IN
#> AGPL                    README.md:129
#>                         README.Rmd:103
#> alf                     intro.Rmd:154
#> amd                     NEWS.md:46
#> aquamaps                aquamapsdata-package.Rd:8
#>                         con_am.Rd:18
#>                         db_minify.Rd:5
#>                         description:1,3
#>                         README.md:69,84
#>                         README.Rmd:75,81
#>                         intro.Rmd:146
#> AquaMaps                am_citation.Rd:16,18,23,24
#>                         am_meta.Rd:18,28,41,47,52
#>                         am_search_exact.Rd:36
#>                         db_minify.Rd:17
#>                         title:1
#>                         README.md:28,81,88,93,100,114
#>                         README.Rmd:41,81,85,88,92,98
#>                         intro.Rmd:12,231,233,235,237,239
#> AquaMaps’             intro.Rmd:231
#> aquamapsdata’         README.md:130
#> aquamapsdb              am_use_offline_db.Rd:12
#>                         aquamapsdata-package.Rd:12
#> au                      csquare_from_dd.Rd:22
#>                         csquare_to_dd.Rd:17
#> Barile                  README.md:81
#>                         README.Rmd:81
#> bathydemersal           am_search_exact.Rd:66
#> bathypelagic            am_search_exact.Rd:66
#> biologie                am_citation.Rd:32
#>                         README.md:107
#>                         README.Rmd:94
#> Bluespotted             intro.Rmd:98
#> Caranx                  intro.Rmd:131
#> Catalogue               am_search_exact.Rd:37
#> catalogueoflife         am_search_exact.Rd:37
#> cells’                intro.Rmd:231
#> chunkwise               intro.Rmd:294
#> cmar                    csquare_from_dd.Rd:22
#>                         csquare_to_dd.Rd:17
#> CMD                     NEWS.md:10
#>                         README.md:5
#>                         README.Rmd:9
#> Codecov                 README.md:10
#>                         README.Rmd:11
#> csiro                   csquare_from_dd.Rd:22
#>                         csquare_to_dd.Rd:17
#> csq                     csquare_from_dd.Rd:22
#>                         csquare_to_dd.Rd:17
#> csquare                 csquare_from_dd.Rd:5,21
#>                         csquare_to_dd.Rd:5,10,16
#> csquarecode             NEWS.md:22
#> CsquareCode             am_csc_from_extent.Rd:19,22
#>                         intro.Rmd:178,208,210
#> CsquareCodes            am_species_in_csc.Rd:10,16
#>                         am_species_per_csc.Rd:10
#> csquares                csquare_from_dd.Rd:22
#>                         csquare_to_dd.Rd:17
#> csv                     NEWS.md:42
#> datadir                 intro.Rmd:259
#> de                      am_citation.Rd:31,32
#>                         README.md:107
#>                         README.Rmd:94
#> degress                 csquare_from_dd.Rd:14
#> DepthPrefMin            am_search_exact.Rd:71
#> dplyr                   aquamapsdata-package.Rd:14
#> duckdb                  con_am.Rd:10
#>                         default_db.Rd:10
#> eval                    NEWS.md:16
#> extdata                 am_use_offline_db.Rd:5,12
#>                         con_am.Rd:10
#>                         default_db.Rd:10,17
#>                         NEWS.md:32
#> faq                     csquare_from_dd.Rd:22
#>                         csquare_to_dd.Rd:17
#> FB                      am_search_exact.Rd:94
#> FC                      am_map_leaflet.Rd:19
#> fcn                     am_custom_query.Rd:14
#> Fis                     am_search_exact.Rd:38
#> FishBase                am_search_exact.Rd:40,46,53,66,69,75,80,82,86,94
#> freiburg                am_citation.Rd:32
#>                         README.md:107
#>                         README.Rmd:94
#> Froese                  am_citation.Rd:31
#>                         README.md:81,107
#>                         README.Rmd:81,94
#>                         intro.Rmd:239
#> fts                     am_search_fuzzy.Rd:11
#> FTS                     NEWS.md:28
#>                         intro.Rmd:302
#> GameFish                am_search_exact.Rd:68
#> geomar                  am_citation.Rd:31
#>                         README.md:107
#>                         README.Rmd:94
#> github                  README.md:36,39
#>                         README.Rmd:45,47
#> gz                      NEWS.md:46
#> hcaf                    am_meta.Rd:38,50
#>                         NEWS.md:30
#>                         intro.Rmd:210,233
#> HCAF                    intro.Rmd:158
#> HSPEC                   am_meta.Rd:39
#> hspen                   am_meta.Rd:47
#>                         NEWS.md:30
#>                         intro.Rmd:190
#> HSPEN                   am_search_exact.Rd:72
#>                         intro.Rmd:192
#> htm                     csquare_from_dd.Rd:22
#>                         csquare_to_dd.Rd:17
#> http                    am_search_fuzzy.Rd:11
#>                         csquare_from_dd.Rd:22
#>                         csquare_to_dd.Rd:17
#> https                   aquamapsdata-package.Rd:12
#> ie                      am_species_per_csc.Rd:5
#> ile                     intro.Rmd:154
#> IPCC                    am_meta.Rd:54
#> IUCN                    am_meta.Rd:36
#>                         am_search_exact.Rd:88,90,92
#> Kaschner                am_citation.Rd:32
#>                         README.md:81,107
#>                         README.Rmd:81,94
#>                         intro.Rmd:239
#> Kesner                  README.md:81
#>                         README.Rmd:81
#> LME                     intro.Rmd:210
#> LOICZID                 which_cells_in_raster.Rd:5,19
#> loiczids                which_cells_in_raster.Rd:16
#> lon                     csquare_to_dd.Rd:13
#> md                      am_citation.Rd:10
#>                         intro.Rmd:120
#> minified                am_use_offline_db.Rd:12
#>                         db_minify_path.Rd:5,10
#>                         default_db.Rd:17
#>                         intro.Rmd:12
#> Minify                  db_minify.Rd:5
#> mysql                   con_am.Rd:10
#>                         db_sync.Rd:35
#>                         default_db.Rd:10
#> NonCommercial           README.md:90
#>                         README.Rmd:86
#> occ                     NEWS.md:30
#> occuring                am_species_in_csc.Rd:5
#>                         intro.Rmd:178
#> occurrencecells         am_meta.Rd:56
#> pkgdown                 NEWS.md:12
#> pre                     con_am.Rd:17
#>                         NEWS.md:44
#> programmatically        README.md:31
#>                         README.Rmd:43
#> Rainer                  am_citation.Rd:31
#>                         README.md:107
#>                         README.Rmd:94
#>                         intro.Rmd:239
#> RCP                     am_meta.Rd:54,55
#>                         intro.Rmd:235
#> Rees                    README.md:81
#>                         README.Rmd:81
#> rfroese                 am_citation.Rd:31
#>                         README.md:107
#>                         README.Rmd:94
#> Rius                    README.md:81
#>                         README.Rmd:81
#> Rmarkdown               intro.Rmd:120
#> ROpenSci                README.md:135
#>                         README.Rmd:106
#> RSQLite                 NEWS.md:16
#>                         README.md:63
#>                         README.Rmd:71
#> Salmo                   NEWS.md:16
#> SeaLifeBase             am_search_exact.Rd:40,46,53,66,75,83,86,94
#> Segschneider            README.md:81
#>                         README.Rmd:81
#> SLB                     am_search_exact.Rd:94
#> species’environmental intro.Rmd:190
#> SpeciesID               am_species_per_csc.Rd:15
#> speciesoccursum         am_meta.Rd:32
#> sql                     am_custom_query.Rd:14
#> sqlite                  am_custom_query.Rd:12
#>                         am_db_sqlite.Rd:5,13
#>                         am_search_fuzzy.Rd:11
#>                         con_am.Rd:10,11
#>                         db_minify_path.Rd:5,10
#>                         default_db.Rd:10,17
#>                         NEWS.md:42
#> stringi                 NEWS.md:24
#> stringr                 NEWS.md:24
#> tibble                  am_csc_from_extent.Rd:19
#>                         am_search_exact.Rd:97
#>                         am_search_fuzzy.Rd:14
#> tidyverse               README.md:135
#>                         README.Rmd:106
#>                         intro.Rmd:42
#> traumatogenic           am_search_exact.Rd:74
#> travis                  NEWS.md:38
#> trevally                intro.Rmd:98
#> trutta                  NEWS.md:16
#> Unported                README.md:90
[…]
