### 2021-01-26 ed2f2af
fix download_db() to use correct unzip wrt has_pass
