### 2021-01-26 d11ae3f
add mentions for other useful packages, fix #29
