### 2021-01-28 1247202
fix spell checks and urls, issue #24 and issue #25
