### 2021-01-28 d0de57a
add vignette answering some FAQs
