### 2021-02-19 #421 @mskyttner
Hi @maurolepore thanks for checking up  on me, I have been off for a while due to some health issues with my elderly parents but getting back now. I have about half of the editor check issues raised adressed (I think) here https://github.com/raquamaps/aquamapsdata/pull/32 but probably need another week to work through the CI / test coverage / style. I tried to use a rocker/geospatial base image rather than rocker/tidyverse base in the CI setup due to some issues with rgdal installation etc, still struggling a bit with getting that sorted out for all the platforms and versions of R etc. Will try to look more in depth at links with examples how this can be set up ^^.
