### 2021-04-30 #421 @maurolepore
@mskyttner I really hope you and your family are well. I realize I touched base a bit too early. The first revision of the 'holding' label will formally be on May 22.

> The holding status will be revisited every 3 months, and after one year the issue will be closed.
> --https://devdevguide.netlify.app/policies.html#policiesreviewprocess
